# .sdd/contracts/api/web-api.openapi.yaml
openapi: 3.1.0
info:
  title: "SDD Web API"
  version: "0.3.0"
  description: |
    REST-API des SDD Web UI Servers (SPEC-0003, CON-0007).
    Läuft lokal auf http://localhost:8000. Kein Auth erforderlich.
  contact:
    name: "Boris"
servers:
  - url: "http://localhost:8000/api"
    description: "Lokaler Dev-Server"
paths:
...
  /projects:
    get:
      operationId: listProjects
      summary: "Alle Projekte auflisten"
      responses:
        "200":
          description: "Liste aller Projekte"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      operationId: createProject
      summary: "Neues Projekt anlegen"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
      responses:
        "201":
          description: "Projekt angelegt"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{project_id}:
    get:
      operationId: getProject
      summary: "Einzelnes Projekt abrufen"
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: "Projekt"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{project_id}/level:
    patch:
      operationId: setProjectLevel
      summary: "Autonomy Level eines Projekts setzen (1|2|3|3.5|4)"
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetLevelRequest"
      responses:
        "200":
          description: "Level aktualisiert"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Specs ─────────────────────────────────────────────────────────────────
  /specs:
    get:
      operationId: listSpecs
      summary: "Alle Specs auflisten"
      responses:
        "200":
          description: "Liste aller Specs"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Spec"

    post:
      operationId: createSpec
      summary: "Neue Spec anlegen"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecCreate"
      responses:
        "201":
          description: "Spec angelegt"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedArtifact"
  /specs/{spec_id}:
    get:
      operationId: getSpec
      summary: "Spec mit Body abrufen"
      parameters:
        - $ref: "#/components/parameters/SpecId"
      responses:
        "200":
          description: "Spec inkl. Markdown-Body"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Contracts ─────────────────────────────────────────────────────────────
  /contracts:
    get:
      operationId: listContracts
      summary: "Alle Contracts auflisten"
      responses:
        "200":
          description: "Liste aller Contracts"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"

    post:
      operationId: createContract
      summary: "Neuen Contract anlegen"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractCreate"
      responses:
        "201":
          description: "Contract angelegt"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedArtifact"
...
  /tests:
    get:
      operationId: listTests
      summary: "Alle Tests auflisten"
      responses:
        "200":
          description: "Liste aller Tests"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Test"

    post:
      operationId: createTest
      summary: "Neuen Test anlegen"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestCreate"
      responses:
        "201":
          description: "Test angelegt"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedArtifact"
  /tests/{test_id}:
    get:
      operationId: getTest
      summary: "Test mit Body abrufen"
      parameters:
        - $ref: "#/components/parameters/TestId"
      responses:
        "200":
          description: "Test inkl. Body"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestDetail"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Commands ──────────────────────────────────────────────────────────────
...
components:
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema: { type: string, pattern: "^PRJ-[0-9]{4,}$" }
    SpecId:
      name: spec_id
      in: path
      required: true
      schema: { type: string, pattern: "^SPEC-[0-9]{4,}$" }
...
    TestId:
      name: test_id
      in: path
      required: true
      schema: { type: string, pattern: "^TST-[0-9]{4,}[a-z]?$" }
  schemas:
    Project:
      type: object
      required: [id, name, owner, status, description, created, autonomy_level]
      properties:
        id:             { type: string, example: "PRJ-0001" }
        name:           { type: string }
        owner:          { type: string }
        status:         { type: string, enum: [active, archived, planning] }
        description:    { type: string }
        created:        { type: string, format: date }
        autonomy_level:
          type: number
          enum: [1, 2, 3, 3.5, 4]
          description: "Dark Factory Autonomy Level (SPEC-0004)"
    ProjectCreate:
      type: object
      required: [name]
      properties:
        name:        { type: string }
        owner:       { type: string, default: "" }
        status:      { type: string, enum: [active, archived, planning], default: active }
        description: { type: string, default: "" }
    SetLevelRequest:
      type: object
      required: [level]
      properties:
        level:
          type: number
          enum: [1, 2, 3, 3.5, 4]
    Spec:
      type: object
      required: [id, title, status]
      properties:
        id:         { type: string, example: "SPEC-0001" }
        project:    { type: string }
        title:      { type: string }
        status:     { type: string, enum: [draft, review, approved, implemented, deprecated] }
        priority:   { type: string, enum: [low, medium, high, critical] }
        owner:      { type: string }
        contracts:  { type: array, items: { type: string } }
        tests:      { type: array, items: { type: string } }
        adrs:       { type: array, items: { type: string } }
        depends_on: { type: array, items: { type: string } }
        tags:       { type: array, items: { type: string } }
        version:    { type: string }
        created:    { type: string }
        updated:    { type: string }
        file:       { type: string }
        abs_file:   { type: string }
    SpecDetail:
      allOf:
        - $ref: "#/components/schemas/Spec"
        - type: object
          properties:
            body: { type: string, description: "Markdown-Body ohne Frontmatter" }
    SpecCreate:
      type: object
      required: [title]
      properties:
        title:      { type: string }
        owner:      { type: string, default: "" }
        priority:   { type: string, default: medium }
        project_id: { type: string, default: "" }
    Contract:
      type: object
      required: [id, title, type, format, spec, status]
      properties:
        id:           { type: string, example: "CON-0001" }
        title:        { type: string }
        type:         { type: string, enum: [api, behavior, data, performance] }
        format:       { type: string }
        spec:         { type: string }
        status:       { type: string, enum: [draft, active, deprecated] }
        version:      { type: string }
        artifact:     { type: string }
        abs_artifact: { type: string }
        tests:        { type: array, items: { type: string } }
        file:         { type: string }
        abs_file:     { type: string }
...
    ContractCreate:
      type: object
      required: [spec_id, format]
      properties:
        spec_id: { type: string }
        format:  { type: string }
        title:   { type: string, default: "" }
    Test:
      type: object
      required: [id, title, level, spec, contract, status]
      properties:
        id:        { type: string, example: "TST-0001" }
        title:     { type: string }
        level:     { type: string, enum: [unit, integration, contract, acceptance, performance, property] }
        spec:      { type: string }
        contract:  { type: string }
        status:    { type: string, enum: [planned, implemented, passing, failing, skipped] }
        framework: { type: string }
        file:      { type: string }
        abs_file:  { type: string }
    TestDetail:
      allOf:
        - $ref: "#/components/schemas/Test"
        - type: object
          properties:
            body: { type: string }
    TestCreate:
      type: object
      required: [spec_id, contract_id, level]
      properties:
        spec_id:     { type: string }
        contract_id: { type: string }
        level:       { type: string, enum: [unit, integration, contract, acceptance, performance, property] }
        title:       { type: string, default: "" }
    CreatedArtifact:
      type: object
      required: [id, file]
      properties:
        id:       { type: string }
        file:     { type: string }
        abs_file: { type: string }
...
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
  responses:
    NotFound:
      description: "Nicht gefunden"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: "Ungültiger Request"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...